Changed Modulus command types
* Changed from one Modulus to Modulus int and modulus float. This prevents some random issues with having different variable types. Alas, you cannot modulo a int and a float together.
