+++ Assets/Fungus/Scripts/Commands/Math/ModulusInt.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 36d63537f4c1a0449b3a102958beb138
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
